+++ config/prod.exs
@@ -19,7 +19,9 @@ config :td_bg, TdBgWeb.Endpoint,
   server: true
 
 # Do not print debug messages in production
-config :logger, :console, format: "[$level] $message\n"
+config :logger, :console,
+  format: (System.get_env("EX_LOGGER_FORMAT") || "[$level] $message") || "\n",
+  level: :info
 
 # ## SSL Support
 #